refine with new added utils.ts
