+++ lib/async-readable.ts
@@ -1,6 +1,5 @@
-/// <reference lib="es2018.asynciterable" />
-
 import { Readable } from 'stream';
+import { reader, rejection } from './utils';
 
 
 
@@ -39,50 +38,6 @@ export function toReadableStream <T> (gen: Gen<T>) {
             objectMode: true,
             read: reader(toAsyncIterable(gen)(source)),
         });
-
-    };
-
-}
-
-
-
-export function reader <T> (source: AsyncIterable<T>, destroy = () => { }) {
-
-    const iterator = source[Symbol.asyncIterator]();
-
-    let reading = false;
-
-    return async function (this: Readable) {
-
-        if (reading) {
-            return;
-        }
-
-        reading = true;
-
-        try {
-
-            while (true) {
-
-                const { value, done } = await iterator.next();
-
-                if (done) {
-                    break;
-                }
-
-                if (this.push(value) === false) {
-                    reading = false;
-                    return;
-                }
-
-            }
-
-            this.push(null);
-
-        } catch (error) {
-            destroy();
-            this.destroy(error);
-        }
 
     };
 
@@ -155,17 +110,6 @@ export function asyncReadable <T extends Buffer> (stream: ReadableStream) {
             next = 0;
             resolve(data);
         }
-
-    }
-
-    function rejection () {
-
-        /* istanbul ignore next */ // tslint:disable-next-line:no-unused-expression
-        let reject = (error: Error) => { error; };
-
-        const error = new Promise((_res, rej) => reject = rej);
-
-        return [ error, reject ] as const;
 
     }
