+++ lib/index.ts
@@ -1,2 +1,3 @@
 export * from './async-readable';
+export { reader } from './utils';
